[POS][HinsChiu]implementation:refactoring with Stream

+++ src/main/java/pos/machine/ReceiptItemInfo.java
@@ -1,14 +1,14 @@
 package pos.machine;
 
 public class ReceiptItemInfo extends ItemInfo{
-    private final int quantity;
+    private final long quantity;
 
-    public ReceiptItemInfo(String barcode, String name, int unitPrice, int quantity) {
+    public ReceiptItemInfo(String barcode, String name, int unitPrice, long quantity) {
         super(barcode, name, unitPrice);
         this.quantity = quantity;
     }
 
-    public int getQuantity() { return quantity; }
+    public long getQuantity() { return quantity; }
 
-    public int getSubTotal() { return super.getPrice() * quantity; }
+    public long getSubTotal() { return super.getPrice() * quantity; }
 }
